modules/auth/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Auth0 JWT Validation
python-jose[cryptography]==3.3.0
httpx==0.25.2

# Config
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Cache
cachetools==5.3.2

